# PERF: optimize local crate index parsing