@vlad20012 2021-07-02
About parallelization: `parallelStream` implicitly uses the global ForkJoinPool instance. There are several issues with it:
1. It is not intended to be used in IO operations.
2. In the IntelliJ platform it is heavily used in highlighting and other stuff. For example, all inspections run from the global ForkJoinPool. This means that heavy or IO tasks submitted to the pool may freeze the highlighting.
3. Tasks that execute on ForkJoinPool may interfere. This is a complex topic... Let's say that during the execution of a task ForkJoinPool may start to execute another task. Of course, it can't do this at a random point, but it can do this when interacting with ForkJoinPool API, for e.g. `ForkJoinTask.get`. This actually means that your code inside `.map {}` may be executed inside a read action, because the platform actually invokes `ForkJoinTask.get` inside a read action. That, in turn, may hang the IDE because of heavy operation inside a read action without cancellation checks =/

In short, we must not use the global ForkJoinPool instance.
I this case it's better to run a new ForkJoinPool and execute the stream on it.
```
val pool = Executors.newWorkStealingPool()
val future = supplyAsync(pool) {
    objectIds
        .parallelStream()
        .map { ... }
        .forEach { ... }        
}
try {
    future.join()
} finally {
    pool.shutdownNow()
}
```
You can also play with a worker thread count. I expect that performance will not increase after some amount of threads, say, 2 or 3.