@Kobzol 2021-07-02
I suspected that the global pool is not a good idea, but I had some trouble getting the local pool to work, I'm not very familiar with the API and I didn't find many examples in Kotlin. I'll check how the performance changes with varying thread counts.