@Kobzol 2021-07-02
Work stealing pool:
- 1 thread: ~23s
- 2 threads: ~15s
- 3 threads: ~13s
- 4 threads: ~12s

I also wanted to try the fixed thread pool, since work stealing often has higher overhead than just simple scheduling. But if I use `Executors.newFixedThreadPool(<any number>)`, it seems to run for the same time and use all of the available cores :man_shrugging:

Iterating the git tree takes about 4 seconds and storing the results into the persistent hash map takes about 3 seconds. Therefore the actual file and JSON parsing part is only about a third of the total computation time and therefore there is probably no point in increasing the number of threads further.

I hope that `repository.open(objectId)` is thread-safe :)